refactor(cli): split scan --baseline/--estimate into cli_scan_baseline
cli_scan.py had grown to 1898/2000 lines (95% of the AI-readiness hard
cap). Extract the two self-contained sub-flows — scan --baseline
(_run_baseline_compare + _baseline_is_native_library) and scan
--estimate (_emit_estimate), plus the shared header-provenance and
--risk-rules helpers — into a new cli_scan_baseline.py sibling, and move
the pure snapshot→coverage-row helpers into the existing
cli_scan_helpers.py. Also drop the dead _uses_fast_binary_surface helper
(no callers anywhere) and relocate _uses_debug_presence_only alongside
the other depth predicates.

cli_scan.py drops 1898 -> 1483 lines (clears the 1500 WARN threshold);
historical import paths (abicheck.cli_scan._run_baseline_compare etc.)
are preserved via re-export for the scan tests and service_scan. Extends
the import-cycle allowlist and the cli_scan mypy override to the new
module. No behaviour change.
