+++ abicheck/cli_scan.py
@@ -61,7 +61,7 @@ from .buildsource.crosscheck import ALL_CHECKS, CrosscheckConfig, run_crosscheck
 from .buildsource.pattern_scan import scan_files
 from .buildsource.poi import build_points_of_interest, resolve_symbol_tus
 from .buildsource.preprocessor_scan import run_preprocessor_scan
-from .buildsource.risk import RiskRules, RiskScore, score_changed_paths
+from .buildsource.risk import RiskScore, score_changed_paths
 from .buildsource.scan_levels import (
     EvidenceDepth,
     ScanMode,
@@ -79,7 +79,20 @@ from .cli_options import (
     verbose_option,
 )
 from .cli_params import DEPTH_PARAM
+from .cli_scan_baseline import (
+    _baseline_is_native_library,  # noqa: F401 - re-export for scan tests/service_scan
+    _emit_estimate,
+    _expand_public_headers,
+    _load_risk_rules,
+    _public_provenance_set,
+    _run_baseline_compare,
+)
 from .cli_scan_helpers import (
+    _intrinsic_coverage,
+    _l3_collected,
+    _pack_coverage,
+    _source_abi_coverage,
+    _uses_debug_presence_only,
     l4_coverage_advisories,
     render_baseline_lines,
     render_coverage_lines,
@@ -285,102 +298,6 @@ class ScanOutcome:
             "elapsed_s": round(self.elapsed_s, 3),
             "budget_s": self.budget_s,
         }
-
-
-def _intrinsic_coverage(snap: Any) -> list[dict[str, Any]]:
-    """Compute the intrinsic L0/L1/L2 coverage rows from a snapshot."""
-    rows: list[dict[str, Any]] = []
-    has_binary = bool(snap.elf or snap.pe or snap.macho)
-    rows.append(
-        {
-            "layer": "L0_binary",
-            "status": "present" if has_binary else "not_collected",
-            "detail": f"{len(snap.functions)} function(s), "
-            f"{len(snap.variables)} variable(s)"
-            if has_binary
-            else "no binary export table (snapshot-only input)",
-        }
-    )
-    dwarf = getattr(snap, "dwarf", None)
-    has_debug = bool(getattr(dwarf, "has_dwarf", False)) if dwarf is not None else False
-    rows.append(
-        {
-            "layer": "L1_debug",
-            "status": "present" if has_debug else "not_collected",
-            "detail": "DWARF/PDB debug info present" if has_debug else "no debug info",
-        }
-    )
-    rows.append(
-        {
-            "layer": "L2_header",
-            "status": "present" if snap.from_headers else "skipped",
-            "detail": f"{len(snap.types)} type(s) from public headers"
-            if snap.from_headers
-            else "no public-header AST (pass --headers; needs castxml or clang)",
-        }
-    )
-    return rows
-
-
-def _source_abi_coverage(snap: Any) -> dict[str, Any]:
-    """Return the embedded L4 coverage dict, if present."""
-    pack = getattr(snap, "build_source", None)
-    surface = getattr(pack, "source_abi", None) if pack is not None else None
-    cov = getattr(surface, "coverage", None) if surface is not None else None
-    return dict(cov or {})
-
-
-def _pack_coverage(snap: Any) -> list[dict[str, Any]]:
-    """Read the L3/L4/L5 coverage rows from a snapshot's embedded pack, if any."""
-    pack = getattr(snap, "build_source", None)
-    if pack is None:
-        return [
-            {
-                "layer": layer,
-                "status": "not_collected",
-                "detail": "no build/source evidence collected "
-                "(pass --sources, or a deeper --source-method)",
-            }
-            for layer in ("L3_build", "L4_source_abi", "L5_source_graph")
-        ]
-    return [c.to_dict() for c in pack.manifest.coverage]
-
-
-def _l3_collected(snap: Any) -> bool:
-    """True when the snapshot carries a non-empty L3 build-evidence layer.
-
-    Used to decide whether a deep ``--source-method`` actually reached L3: a
-    ``not_collected`` (or absent pack) L3 means the requested L3/L4/L5 layers were
-    skipped for want of a compile database, which warrants a pointed advisory.
-    ``partial`` counts as collected — it ran and produced something.
-    """
-    pack = getattr(snap, "build_source", None)
-    if pack is None:
-        return False
-    for cov in pack.manifest.coverage:
-        row = cov.to_dict() if hasattr(cov, "to_dict") else cov
-        if row.get("layer") == "L3_build":
-            return row.get("status") != "not_collected"
-    return False
-
-
-def _uses_fast_binary_surface(depth: EvidenceDepth) -> bool:
-    """True when the scan depth needs only ELF exports plus cheap debug presence.
-
-    The deeper DWARF DIE walk is source/type evidence. ``headers`` gets its type
-    evidence from L2 AST, and ``build`` adds L3 compile context; neither needs the
-    expensive DWARF expansion on the binary side.
-    """
-    return depth in {
-        EvidenceDepth.BINARY,
-        EvidenceDepth.HEADERS,
-        EvidenceDepth.BUILD,
-    }
-
-
-def _uses_debug_presence_only(depth: EvidenceDepth) -> bool:
-    """True when L2/L3 evidence is collected elsewhere, so DWARF stays cheap."""
-    return depth in {EvidenceDepth.HEADERS, EvidenceDepth.BUILD}
 
 
 def _normalize_depth_inputs(
@@ -1564,335 +1481,3 @@ def run_scan_core(
         outcome=outcome, findings=list(cc.findings), snapshot=new_snap
     )
 
-
-def _public_provenance_set(
-    headers: list[Path], public_header_dirs: list[Path]
-) -> tuple[list[Path], list[Path]]:
-    """Build the ``(public_headers, public_header_dirs)`` provenance set for scan.
-
-    A directory boundary is what lets ``apply_provenance`` classify origins as
-    PUBLIC/INTERNAL (and so unlocks the leakage / RTTI / exported-vs-public
-    cross-checks, ADR-024). Directories come from ``--public-header-dir`` and from
-    any ``-H`` argument that is itself a directory; ``-H`` *file* arguments ride
-    along as explicit public headers.
-
-    A lone ``-H`` umbrella *file* with no directory does **not** activate
-    provenance: a single header cannot establish a public directory boundary
-    (the abicheck A1 finding), so we return empty sets and every origin stays
-    ``UNKNOWN`` — preserving the prior default-scan behaviour.
-    """
-    dirs = list(public_header_dirs)
-    files: list[Path] = []
-    for h in headers:
-        if h.is_dir():
-            dirs.append(h)
-        else:
-            files.append(h)
-    if not dirs:
-        return [], []
-    return files, dirs
-
-
-def _expand_public_headers(headers: list[Path]) -> list[str]:
-    """Expand ``-H`` inputs (files or directories) to individual header files.
-
-    ``-H/--headers`` accepts a directory (the snapshot build expands it the same
-    way); the S2 leak pass needs the individual header *files* so clang
-    preprocesses each one, not a directory as a single bogus TU. Falls back to the
-    raw paths if expansion fails (e.g. an empty dir) so the pass still runs.
-    """
-    from .service import expand_header_inputs
-
-    try:
-        return [str(p) for p in expand_header_inputs(headers)]
-    except Exception:  # noqa: BLE001 - expansion is best-effort for the advisory tier
-        return [str(h) for h in headers]
-
-
-def _emit_estimate(
-    *,
-    binary: Path,
-    headers: list[Path],
-    includes: list[Path],
-    sources: Path | None,
-    build_info: Path | None,
-    mode: str,
-    resolved_method: SourceMethod,
-    eff_depth: EvidenceDepth,
-    changed: list[str],
-    seeded: bool,
-    budget_s: float | None,
-    lang: str,
-    fmt: str,
-    output: Path | None,
-) -> None:
-    """Render the ADR-035 D10 dry-run cost estimate (``scan --estimate``).
-
-    A thin front-end over :func:`service.estimate_scan`: builds a
-    :class:`service.ScanRequest`, probes the project (TU count, header fan-out)
-    and prints the projected per-layer cost — scanning nothing, running no
-    compiler. Always exits 0 (it is a probe, not a gate).
-    """
-    from .service import Budget, ScanRequest, estimate_scan
-
-    req = ScanRequest(
-        binaries=[binary],
-        headers=headers,
-        includes=includes,
-        sources=sources,
-        build_info=build_info,
-        mode=mode,
-        source_method=resolved_method.value,
-        depth=eff_depth.value,
-        changed_paths=list(changed),
-        seeded=seeded,
-        budget=Budget(total_timeout=budget_s),
-        lang=lang,
-    )
-    # Pass the *already-resolved* level so the estimate mirrors the real scan
-    # exactly — re-resolving from the round-tripped flags would re-apply the
-    # source-method > depth precedence and lose a mode preset's deeper depth
-    # (pr-deep = (s5, graph)); Codex review.
-    estimates = estimate_scan(req, resolved_level=(resolved_method, eff_depth))
-    total = sum(e.est_seconds for e in estimates)
-
-    if fmt == "json":
-        text = json.dumps(
-            {
-                "mode": mode,
-                "estimate": [e.to_dict() for e in estimates],
-                "total_est_seconds": round(total, 3),
-            },
-            indent=2,
-        )
-    else:
-        lines = [
-            f"abicheck scan --estimate — {mode} mode (dry run; nothing scanned)",
-            "",
-        ]
-        lines.append(f"  {'layer':<16} {'method':<8} {'TUs':>6}  {'est_s':>8}  note")
-        for e in estimates:
-            lines.append(
-                f"  {e.layer:<16} {(e.method or '-'):<8} {e.tus:>6}  "
-                f"{e.est_seconds:>8.2f}  {e.note}"
-            )
-        lines.append("")
-        lines.append(f"  projected total: {total:.2f}s")
-        text = "\n".join(lines)
-
-    if output:
-        _safe_write_output(output, text)
-        click.echo(f"Estimate written to {output}", err=True)
-    else:
-        click.echo(text)
-
-
-def _load_risk_rules(path: Path | None) -> RiskRules:
-    """Load a ``risk_rules:`` profile from a YAML file, or the shipped default."""
-    if path is None:
-        return RiskRules.default()
-    import yaml  # hard dep (pyyaml); import out of the try so the except can name it
-
-    try:
-        raw = yaml.safe_load(path.read_text(encoding="utf-8"))
-    except (OSError, ValueError, yaml.YAMLError) as exc:
-        # yaml.YAMLError (e.g. ParserError) is not a ValueError, so catch it
-        # explicitly — else malformed --risk-rules YAML escapes as a traceback
-        # through the installed console script (Codex review).
-        raise click.ClickException(f"cannot read --risk-rules {path}: {exc}") from exc
-    block = raw.get("risk_rules") if isinstance(raw, dict) else None
-    return RiskRules.from_dict(block if isinstance(block, dict) else raw)
-
-
-def _baseline_is_native_library(path: Path) -> bool:
-    """True if *path* is a native binary, not a JSON / ABICC-dump snapshot.
-
-    A snapshot baseline already has its headers baked in, so the candidate-`-H`
-    reuse is harmless there; only a native binary is re-parsed (and thus at risk
-    of being read through the wrong headers).
-
-    Detection is content-first to match `resolve_input`'s own native dispatch:
-    magic-byte sniffing (`detect_binary_format`) catches the cases a suffix scan
-    misses — an extensionless ELF (`build/foo`), a Mach-O framework binary, a
-    `.pyd`/`.node` shared object (Codex review). The filename heuristic is only a
-    fallback for paths that cannot be sniffed (e.g. a not-yet-existing file in a
-    unit test), and the snapshot suffixes short-circuit first so a real `.json`
-    on disk is never mis-sniffed.
-    """
-    name = path.name.lower()
-    if name.endswith((".json", ".dump", ".tar.gz", ".tgz", ".xml")):
-        return False
-    from .binary_utils import detect_binary_format
-
-    if detect_binary_format(path) is not None:
-        return True
-    return ".so" in name or name.endswith((".dll", ".dylib"))
-
-
-def _run_baseline_compare(
-    baseline: Path,
-    binary: Path,
-    new_snap: Any,
-    extra_changes: list[Any],
-    lang: str,
-    collect_mode: str,
-    headers: list[Path],
-    includes: list[Path],
-    public_headers: list[Path],
-    public_header_dirs: list[Path],
-    compile_context: CompileContext | None = None,
-    baseline_headers: list[Path] | None = None,
-    baseline_includes: list[Path] | None = None,
-    symbols_only: bool = False,
-    debug_presence_only: bool = False,
-) -> tuple[str, int, dict[str, Any]]:
-    """Compare *new_snap* against *baseline*, preserving scan authority.
-
-    Single-version cross-source findings are reported in the scan's dedicated
-    ``crosscheck`` block and stay advisory for baseline comparisons unless the
-    maintainer explicitly promotes one with ``--crosscheck KEY=error``. They are
-    not folded into ``extra_changes`` by default: doing so lets a candidate-side
-    evidence hygiene finding such as ``header_build_context_mismatch`` turn a
-    clean old/new artifact diff into an ``API_BREAK`` false positive. Real
-    old/new embedded build/source drift is still diffed below via
-    ``prepare_embedded_build_source``.
-
-    *headers*/*includes* are the same scan header inputs used to build the
-    candidate, threaded into the baseline parse so a native ``--baseline``
-    library is header-scoped symmetrically — else the old side stays
-    symbol/DWARF-only and the compare drops old type evidence or invents spurious
-    API diffs (Codex review). They are inert for a JSON-snapshot baseline.
-
-    The embedded L3/L4/L5 build/source packs on either snapshot are diffed via
-    :func:`prepare_embedded_build_source` — the same path ``abicheck compare``
-    uses — so source-only / graph findings the collected evidence reveals are
-    folded into the verdict too (``checker.compare`` itself does not read
-    ``build_source``).
-    """
-    from .cli_buildsource import prepare_embedded_build_source
-    from .errors import AbicheckError
-    from .service import compare_snapshots, resolve_input
-
-    # Each side is parsed with its *own* headers. `scan` has a single -H (built for
-    # the candidate); for a native --baseline library whose public headers differ,
-    # --baseline-header/-include select the old side's headers. Without them we
-    # reuse the candidate -H/-I — correct only when the headers did not change — so
-    # warn rather than silently read the old side through the new headers (Codex).
-    if baseline_headers:
-        bl_headers = list(baseline_headers)
-        bl_includes = list(baseline_includes) if baseline_includes else includes
-        bl_public_headers = bl_headers
-        # The old-side public boundary comes ONLY from --baseline-header: dirs in
-        # it are public-header dirs, files opt in just themselves. Do NOT fall back
-        # to the new side's public dirs — a relative dir like `include/` would
-        # (segment-based provenance) re-mark old private headers as PUBLIC and skew
-        # the public-surface scoping (Codex review).
-        bl_public_dirs = [p for p in bl_headers if p.is_dir()]
-    else:
-        bl_headers, bl_includes = headers, includes
-        bl_public_headers, bl_public_dirs = public_headers, public_header_dirs
-        if headers and _baseline_is_native_library(baseline):
-            click.echo(
-                f"warning: --baseline {baseline.name} is a native library parsed "
-                f"with the new build's headers (-H); if its public headers differ "
-                f"from the new version, pass --baseline-header (else the old side is "
-                f"read through the new headers and the diff may be wrong/noisy).",
-                err=True,
-            )
-
-    try:
-        old_snap = resolve_input(
-            baseline,
-            bl_headers,
-            bl_includes,
-            version="",
-            lang=lang,
-            public_headers=bl_public_headers,
-            public_header_dirs=bl_public_dirs,
-            compile=compile_context,
-            symbols_only=symbols_only,
-            debug_presence_only=debug_presence_only,
-        )
-    except AbicheckError as exc:
-        raise click.ClickException(
-            f"Failed to load --baseline {baseline}: {exc}"
-        ) from exc
-
-    # Preserve hard L0 removals even when the richer header/source view cannot
-    # prove public-header ownership for the removed entity.  A source/full scan
-    # may parse both sides' headers through different consumer macro contexts;
-    # in fixtures such as case97 the old library exported a function that the
-    # old header exposes only under a consumer macro, so the final public-surface
-    # comparison can otherwise filter the old-only ELF fact away.  Re-reading the
-    # already-loaded snapshots without public-surface scoping and carrying only
-    # the hard ELF-only removal kind keeps the L0 authority while avoiding the
-    # older false-positive class where advisory cross-check findings were folded
-    # into the verdict wholesale.
-    l0_hard_removals: list[Any] = []
-    if not symbols_only:
-        l0_old_snap = resolve_input(
-            baseline,
-            [],
-            [],
-            version="",
-            lang=lang,
-            symbols_only=True,
-        )
-        l0_new_snap = resolve_input(
-            binary,
-            [],
-            [],
-            version="",
-            lang=lang,
-            symbols_only=True,
-        )
-        l0_diff = compare_snapshots(
-            l0_old_snap,
-            l0_new_snap,
-            extra_changes=[],
-            scope_to_public_surface=False,
-        )
-        l0_hard_removals = [
-            change
-            for change in getattr(l0_diff, "breaking", ())
-            if getattr(getattr(change, "kind", None), "value", None)
-            == "func_removed_elf_only"
-        ]
-    # Fold embedded build-info/source (L3/L4/L5) diff findings into extra_changes
-    # before comparing — mirrors the compare command (Codex review). Only engage
-    # when a snapshot actually carries an embedded pack; otherwise pass
-    # ``collect_mode="off"`` so the pipeline stays inert (no spurious collection
-    # attempt / output noise on a plain artifact-only baseline compare).
-    has_embedded = (
-        old_snap.build_source is not None or new_snap.build_source is not None
-    )
-    merged_extra, _coverage_rows, _metrics, _ev = prepare_embedded_build_source(
-        old_snap,
-        new_snap,
-        collect_mode if has_embedded else "off",
-        [*extra_changes, *l0_hard_removals],
-        None,
-        None,
-        None,
-        None,
-    )
-    diff = compare_snapshots(
-        old_snap,
-        new_snap,
-        extra_changes=merged_extra,
-        scope_to_public_surface=True,
-    )
-    summary = {
-        "breaking": len(diff.breaking),
-        "api_break": len(diff.source_breaks),
-        "risk": len(diff.risk),
-        "compatible": len(diff.compatible),
-    }
-    verdict = diff.verdict.value
-    if verdict == "BREAKING":
-        exit_code = 4
-    elif verdict == "API_BREAK":
-        exit_code = 2
-    else:
-        exit_code = 0
-    return verdict, exit_code, summary
